recoil persist 추가

See merge request s08-final/S08P31C101!262

+++ backend/user-service/src/main/java/com/example/userservice/dto/response/user/LocationResponseDto.java
@@ -13,8 +13,8 @@ import lombok.NoArgsConstructor;
 public class LocationResponseDto {
 
     private Long userId;
-    private double lat;
-    private double lng;
+    private Float lat;
+    private Float lng;
 
     public static LocationResponseDto from(Location location) {
         return LocationResponseDto.builder()
